Move Durable Object files to src/storage and update imports
